requirements.txt
awscli
boto3
streamlit
langchain
pypdf
faiss-cpu
langchain_community
python-dotenv